% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_description}
\alias{add_description}
\title{Add a BC GOV standard DESCRIPTION file to the project directory}
\usage{
add_description(path = ".")
}
\arguments{
\item{path}{Directory path (default \code{"."})}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Add a BC GOV standard DESCRIPTION file to the project directory
}
\seealso{
\code{\link{add_readme}}, \code{\link{add_license}}, \code{\link{add_license_header}}
}
