% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_license}
\alias{add_license}
\title{Add a LICENSE file (Apache 2.0) to the project directory}
\usage{
add_license(path = ".", package_desc = FALSE)
}
\arguments{
\item{path}{Directory path (default \code{"."})}

\item{package_desc}{Should the license be added to the DESCRIPTION file if this is a package?}
}
\description{
Add a LICENSE file (Apache 2.0) to the project directory
}
\seealso{
\code{\link{add_readme}}, \code{\link{add_contributing}}, \code{\link{add_license_header}}
}

