% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_license_header}
\alias{add_license_header}
\title{Add the boilerplate Apache header to the top of a source code file}
\usage{
add_license_header(file, year,
  copyright_holder = "Province of British Columbia")
}
\arguments{
\item{file}{Path to the file}

\item{year}{The year the license should apply}

\item{copyright_holder}{Copyright holder (Default "Province of British Columbia")}
}
\description{
Add the boilerplate Apache header to the top of a source code file
}
\seealso{
\code{\link{add_license}}
}

