% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_rbuildignore}
\alias{add_rbuildignore}
\title{Add a RBuildignore file to the project directory}
\usage{
add_rbuildignore(path = ".")
}
\arguments{
\item{path}{Directory path (default \code{"."})}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Add a RBuildignore file to the project directory
}
