% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_readme}
\alias{add_readme}
\title{Add a README.md file to the project directory}
\usage{
add_readme(path = ".", package = FALSE, CoC = TRUE)
}
\arguments{
\item{path}{Directory path (default \code{"."})}

\item{package}{Is this a package or a regular project? (Default \code{FALSE})}

\item{CoC}{Should a Code of Conduct be added to the repository?}
}
\description{
Add a README.md file to the project directory
}
\seealso{
\code{\link{add_contributing}}, \code{\link{add_license}}, \code{\link{add_license_header}}
}
