% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_rproj}
\alias{add_rproj}
\title{Add an RProject file to a directory}
\usage{
add_rproj(path = ".")
}
\arguments{
\item{path}{folder path of the project}
}
\description{
Add an RProject file to a directory
}

