% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_to_rbuildignore}
\alias{add_to_rbuildignore}
\title{Add text to Rbuildignore}
\usage{
add_to_rbuildignore(path, text)
}
\arguments{
\item{path}{}

\item{text}{}
}
\value{
TRUE
}
\description{
Add text to Rbuildignore
}
\keyword{internal}

