% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_skeleton.R
\name{analysis_skeleton}
\alias{analysis_skeleton}
\title{Creates the framework of a new analysis development folder}
\usage{
analysis_skeleton(path = ".", git_init = TRUE, git_clone = NULL,
  rstudio = TRUE, apache = TRUE, CoC = TRUE, rmarkdown = TRUE,
  coc_email = getOption("bcgovr.coc_email", default = NULL),
  dir_struct = getOption("bcgovr.dir.struct", default = NULL),
  copyright_holder = "Province of British Columbia")
}
\arguments{
\item{path}{location to create new analysis. If \code{"."} (the default), 
the name of the working directory will be taken as the analysis name. If 
not \code{"."}, the last component of the given path will be used as the 
analysis name.}

\item{git_init}{Create a new git repository? Logical, default \code{TRUE}.}

\item{git_clone}{the url of a git repo to clone.}

\item{rstudio}{Create an Rstudio project file?}

\item{apache}{Add licensing info for release under Apache 2.0? Default \code{TRUE}.}

\item{CoC}{Should a Code of Conduct be added to the repository? Default \code{TRUE}.}

\item{rmarkdown}{Should an rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{TRUE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}

\item{dir_struct}{alternative analysis directory structure. This should be specified as
a character vector of directory and file paths (relative to the root of the project). 
Directories should be identified as such by having a trailing forward-slash (e.g., \code{"dir/"}).
The default is: \code{c("R/","out/", "graphics/", "data/", "01_load.R", "02_clean.R", "03_analysis.R", "04_output.R", "internal.R", "run_all.R")}.
This can also be set as an option \code{bcgovr.dir.struct}. You may want to set this in your 
.Rprofile file so that every time you start a new analysis, your custom structure is set up.
The line in your \code{.Rprofile} file would look something like this: 
\code{options("bcgovr.dir.struct" = c("doc/", "data/", "bin/", "results/", "src/01_load.R", "src/02_clean.R", "src/03_analysis.r", "src/04_output.R", "src/runall.R"))}}

\item{copyright_holder}{the name of the copyright holder (default 
"Province of British Columbia). Only necessary if adding a license}
}
\description{
Creates the folder structure for a new analysis.
}
\details{
If you are cloning a repository (\code{git_clone = "path_to_repo"}),
  you should run this function from the root of your dev folder and leave 
  \code{path = "."}, as the repository will be cloned into a new folder. If 
  you are setting up a new project (with or without git), \code{path} should 
  be \code{"."} if you are within an already created project directory, or 
  the name of the folder you want to create.
}
\examples{
\donttest{
 bcgovr::analysis_skeleton(path = "c:/_dev/tarballs")
}
}
