% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgovr-package.R
\docType{package}
\name{bcgovr-package}
\alias{bcgovr}
\alias{bcgovr-package}
\title{bcgovr: Set up of bcgov R projects & packages}
\description{
The 'bcgovr' package provides functions and RStudio Addins that automate the creation and set up of bcgov R projects and packages that meet bcgov GitHub requirements.
}
\references{
bcgovr functions:

\itemize{
\item Functions for automating the creation and set up of bcgov R
projects and packages for bcgov GitHub
\item RStudio Addins for inserting licence headers
and BCDevExchange project state badges

}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/bcgovr}
  \item Report bugs at \url{https://github.com/bcgov/bcgovr/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gov.bc.ca}

Authors:
\itemize{
  \item Stephanie Hazlitt \email{stephanie.hazlitt@gov.bc.ca}
  \item Sam Albers \email{sam.albers@gov.bc.ca}
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
