% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgovr.R
\docType{package}
\name{bcgovr}
\alias{bcgovr}
\alias{bcgovr-package}
\title{bcgovr: A package for helping create project directories that meet BC Government GitHub requirements.}
\description{
bcgovr: A package for helping create project directories that meet BC Government GitHub requirements.
}
\section{bcgovr functions}{


\itemize{
\item Functions for auto-populating a new R-based data
analysis or package project directory
\item RStudio addins for adding the Apache 2.0 license header
and BCDevExchange project state badges

}
}

