% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_colours.R
\name{bgc_colours}
\alias{bgc_colours}
\title{Get colour codes for B.C. Biogeoclimatic (BGC) Zones}
\usage{
bgc_colours(which = NULL)
}
\arguments{
\item{which}{Optionally specify the codes for the BGC zones you want. 
If \code{NULL} (the default), will return colours for all zones}
}
\value{
A named character vector of hexadecimal colour codes, where the names are BGC Zone codes
}
\description{
Get colour codes for B.C. Biogeoclimatic (BGC) Zones
}
\examples{
bgc_colours()
bgc_colours(c("BAFA", "CWH"))
}
