% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licence_header.R
\name{check_licence_header}
\alias{check_licence_header}
\title{Check for presence of a licence header (Apache 2.0 or CC-BY) in one or more
files}
\usage{
check_licence_header(files, licence = c("apache2", "cc-by"))
}
\arguments{
\item{licence}{which licence header to check for?
One of \code{c("apache2", "cc-by")}}

\item{file}{path to a file}
}
\value{
named logical vector indicating if each file has the specified
licence header
}
\description{
Check for presence of a licence header (Apache 2.0 or CC-BY) in one or more
files
}
