% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_git.R
\name{clone_git}
\alias{clone_git}
\title{Clones a remote repository}
\usage{
clone_git(url, path)
}
\arguments{
\item{url}{url of remote git repository}

\item{path}{path to clone into}
}
\value{
a repository object
}
\description{
Clones a remote repository
}
