% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_bcgov_package}
\alias{create_bcgov_package}
\title{Creates the framework of a new package development folder}
\usage{
create_bcgov_package(path = ".", git_init = TRUE, git_clone = NULL,
  apache = TRUE, rstudio = TRUE, CoC = TRUE, rmarkdown = TRUE,
  coc_email = getOption("bcgovr.coc.email"),
  copyright_holder = "Province of British Columbia")
}
\arguments{
\item{rmarkdown}{Should an rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{TRUE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}
}
\description{
Creates the framework of a new package development folder
}
\examples{
\donttest{
 bcgovr::create_bcgov_package(path = "c:/_dev/tarballs")
}
}
