% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devex_badge.R
\name{devex_badge}
\alias{devex_badge}
\title{This function is deprecated, use the \code{insert_bcgov_lifecycle_badge} function.}
\usage{
devex_badge(project_state, cat = TRUE)
}
\arguments{
\item{project_state}{One of: 'inspiration', 'exploration', 'dormant', 'delivery' or 'retired'}

\item{cat}{Use cat to print the result (\code{TRUE}; default) or return a character vector (\code{FALSE})?}
}
\value{
html
}
\description{
This function is deprecated, use the \code{insert_bcgov_lifecycle_badge} function.
}
