% Generated by roxygen2 (4.0.2): do not edit by hand
\name{facet_label_wrap}
\alias{facet_label_wrap}
\title{Provides a wrapping function to pass to \code{labeller} argument in \code{ggplot2}'s
\code{facet_grid}}
\usage{
facet_label_wrap(width = 25)
}
\arguments{
\item{width}{The width at which to wrap (in number of characters)}
}
\value{
None
}
\description{
Provides a wrapping function to pass to \code{labeller} argument in \code{ggplot2}'s
\code{facet_grid}
}
\details{
Usage in ggplot2 with facet_grid: \code{ggplot(foo...) + facet_grid(facet_var ~ facet_var2, labeller = facet_label_wrap(25))}
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link[ggplot2]{facet_grid}}
}
\keyword{facet_grid}
\keyword{ggplot2}
\keyword{labeller}
\keyword{wrap}

