% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_license.R
\name{get_data_license}
\alias{get_data_license}
\title{Returns a url, or a markdown or html-formatted link to one of several B.C. or Candadian licenses}
\usage{
get_data_license(license = c("ogl-bc", "bc-crown", "ogl-canada", "statscan"),
  what = "md_link")
}
\arguments{
\item{license}{Which license? One of:
\itemize{
  \item ogl-bc Open Government License - British Columbia
  \item bc-crown B.C. Crown Copyright
  \item ogl-canada Open Government License - Canada
  \item statscan Statistics Canada Open Licence Agreement
}}

\item{what}{What do you want returned? One of
\itemize{
  \item md_link A markdown-formatted link (default)
  \item html_link An HTML-formatted link
  \item url The url of the license
}}
}
\value{
character containing the desired output
}
\description{
Returns a url, or a markdown or html-formatted link to one of several B.C. or Candadian licenses
}
\examples{
get_data_license("ogl-bc", "url")

get_data_license("ogl-canada")

get_data_license("statscan", "html_link")
}
