% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devex_badge.R
\name{insert_bcgov_lifecycle_badge}
\alias{insert_bcgov_lifecycle_badge}
\title{Add html for inserting a bcgov GitHub lifecycle project state badge}
\usage{
insert_bcgov_lifecycle_badge(project_state, cat = TRUE)
}
\arguments{
\item{project_state}{One of: 'experimental', 'maturing', 'dormant', 'stable' or 'retired'}

\item{cat}{Use cat to print the result (\code{TRUE}; default) or return a character vector (\code{FALSE})?}
}
\value{
html
}
\description{
Add html for inserting a bcgov GitHub lifecycle project state badge
}
