% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Plot multiple plots in a single pane}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL, widths = NULL,
  heights = NULL, title = NULL, titlefont = "", titleface = 1,
  titlesize = 16)
}
\arguments{
\item{...}{Two or more ggplot2 objects}

\item{plotlist}{(optional) a list of ggplot2 objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}

\item{widths}{a vector of relative column widths eg. c(3,2)}

\item{heights}{a vector of relative column heights eg. c(3,2)}

\item{title}{Optional title as a character string}

\item{titlefont}{The font of the title}

\item{titleface}{The font face (1 = normal, 2 = bold, 3 = italic, 4 = bold italic)}

\item{titlesize}{The sice of the title font}
}
\value{
Don't know
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.  To save, you must use the desired device (eg \code{png()}),
or save from the RStudio Viewer.

Borrowed and modified from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
\dontrun{
library("ggplot2")
plot1 <- ggplot(iris, aes(x = Species, y = Sepal.Length)) +
   geom_bar(stat = "identity")
plot2 <- ggplot(mtcars, aes(x = mpg, y = disp)) +
   geom_smooth()
multiplot(plot1, plot2, cols = 2, widths = c(3,2), title = "My two unrelated plots")
multiplot(plot1, plot2, cols = 1, heights = c(10,2), title = "My two unrelated plots")
myplots <- list(plot1, plot2, plot1)
multiplot(plotlist = myplots, layout =matrix(c(1,2,3,3), nrow=2),
     heights = c(1,3), widths = c(3,4), title = "My three unrelated plots")
## Adjusting fonts
library(extrafont)
loadfonts("win")
multiplot(plotlist = myplots, layout =matrix(c(1,2,3,3), nrow=2),
          heights = c(1,3), widths = c(3,4), title = "My three unrelated plots",
          titlefont = "Wingdings", titleface = 4, titlesize = 20)
}
}

