% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roxygen_template}
\alias{roxygen_template}
\title{Populate the boilerplate roxygen template at the top of the file containing the function.}
\usage{
roxygen_template(funfile)
}
\arguments{
\item{funfile}{path to the .R file containing the function}
}
\value{
nothing, but adds the roxygen template to the top of the file
}
\description{
Populate the boilerplate roxygen template at the top of the file containing the function.
}
\details{
File must have the argument list and the opening \code{\{} all on one line:
           \code{myFun <- function(arg1, arg2, arg3) \{}
}
\examples{
\dontrun{

}
}

