% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_ghostscript}
\alias{set_ghostscript}
\title{Set the environment variable for the path to the ghostscript executable.}
\usage{
set_ghostscript(gsfile = "C:/Program Files/gs/gs9.15/bin/gswin64c.exe")
}
\arguments{
\item{gsfile}{path to the ghostscript executable}
}
\value{
logical dependent on whether or not the file exists
}
\description{
Set the environment variable for the path to the ghostscript executable.
}
\keyword{internal}

