% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_setup.R
\name{set_home}
\alias{set_home}
\title{Set HOME environment variable on Windows}
\usage{
set_home()
}
\value{
the path to \code{HOME}, invisibly
}
\details{
If you have a \code{.Rprofile} or \code{.Renviron} file in your existing \code{HOME}
directory these will be copied to the new location of \code{HOME}.

This also ensures that the user's personal package library is at
\code{C:/Users/[your_username]/R/win-library/[R-version]/} if you
haven't already set it via the \code{R_LIBS} environment variable. It also moves
all of your packages from your old library location to the new one.
}
\seealso{
set_cran_repo
}
