% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_setup.R
\name{setup_r}
\alias{setup_r}
\title{Set up R for use on a standard B.C. government Windows computer}
\usage{
setup_r()
}
\description{
Set up R for use on a standard B.C. government Windows computer
}
\seealso{
set_home set_cran_repo
}
