% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_soe}
\alias{theme_soe}
\title{Default theme for EnvReportBC graphs and plots}
\usage{
theme_soe(base_size = 12, base_family = "Verdana", use_sizes = TRUE,
  facet = FALSE)
}
\arguments{
\item{base_size}{base font size (default = 12)}

\item{base_family}{base font family (default = Verdana)}

\item{use_sizes}{use relative font sizes (?)}

\item{facet}{facetting? (advise to not use directly, is set by \code{theme_soe_facet})}
}
\value{
returns a plot theme
}
\description{
Default theme for EnvReportBC graphs and plots
}
\examples{
\dontrun{

}
}
\keyword{plotting}
\keyword{theme}

