% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_soe.r
\name{theme_soe_facet}
\alias{theme_soe_facet}
\title{Soe plot theme for facetted graphs}
\usage{
theme_soe_facet(base_size = 12, base_family = "Verdana")
}
\arguments{
\item{base_size}{base font size (default = 12)}

\item{base_family}{base font family (default = Verdana)}
}
\value{
a ggplot2 theme
}
\description{
Soe plot theme for facetted graphs
}
\keyword{plotting}
\keyword{theme}

