% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_code_of_conduct}
\alias{use_bcgov_code_of_conduct}
\title{Add a CODE_OF_CONDUCT.md file to the project directory}
\usage{
use_bcgov_code_of_conduct(package = FALSE,
  coc_email = getOption("bcgovr.coc.email", default = NULL))
}
\arguments{
\item{package}{Is this a package or a regular project? (Default  \code{FALSE})}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Add a CODE_OF_CONDUCT.md file to the project directory
}
\seealso{
\code{\link[=use_bcgov_readme]{use_bcgov_readme()}}, \code{\link[=use_bcgov_licence]{use_bcgov_licence()}}, \code{\link[=use_bcgov_contributing]{use_bcgov_contributing()}}
}
