% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_contributing}
\alias{use_bcgov_contributing}
\title{Add a CONTRIBUTING.md file to the project directory}
\usage{
use_bcgov_contributing(package = FALSE)
}
\arguments{
\item{package}{Is this a package or a regular project? (Default \code{FALSE}).
If \code{TRUE}, "CONTRIBUTING.md" will be added to .Rbuildignore}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Add a CONTRIBUTING.md file to the project directory
}
\seealso{
\code{\link[=use_bcgov_readme]{use_bcgov_readme()}}, \code{\link[=use_bcgov_licence]{use_bcgov_licence()}}, \code{\link[=add_license_header]{add_license_header()}}
}
