% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_git}
\alias{use_bcgov_git}
\title{Initialise a git repository, with bcgov GitHub requirements}
\usage{
use_bcgov_git(rmarkdown = TRUE, licence = "apache2",
  coc_email = get_coc_email(), message = "Initial commit")
}
\arguments{
\item{rmarkdown}{Should an Rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{FALSE}.}

\item{licence}{Which licence to apply? Default is Apache 2.0 (\code{"apache2"}).
Use \code{"cc-by"} for \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0}}

\item{coc_email}{Contact email address(es) for the Code of Conduct. It is
recommended that you save this setting by adding a line like:
\code{options("bcgovr.coc.email" = "your.email@gov.bc.ca")} to yor \code{.Rprofile}.
Use \code{\link[usethis:edit]{usethis::edit_r_profile()}} for an easy way to find and edit this file.}

\item{message}{git commit message}
}
\description{
Initialise a git repository, with bcgov GitHub requirements
}
