% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_gitattributes}
\alias{use_bcgov_gitattributes}
\title{Add an R-flavoured .gitattributes file}
\usage{
use_bcgov_gitattributes()
}
\description{
Adds a .gitattributes file that will identify the repository as an R-based code project
even if many lines of another language are present.
}
