% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_pkgdown}
\alias{use_bcgov_pkgdown}
\title{Set up a \href{https://pkgdown.r-lib.org/index.html}{pkgdown} site using a
bcgov-themed template}
\usage{
use_bcgov_pkgdown(config_file = "_pkgdown.yml", destdir = "docs")
}
\arguments{
\item{config_file}{Path to the pkgdown yaml config file}

\item{destdir}{Target directory for pkgdown docs}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Set up a \href{https://pkgdown.r-lib.org/index.html}{pkgdown} site using a
bcgov-themed template
}
\seealso{
\code{\link[usethis:use_pkgdown]{usethis::use_pkgdown()}} \code{\link[pkgdown:build_site]{pkgdown::build_site()}}
}
