% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_readme}
\alias{use_bcgov_readme}
\title{Add a README.md file to the project directory}
\usage{
use_bcgov_readme(path = ".", project = NULL, package = FALSE)
}
\arguments{
\item{path}{Directory path (default \code{"."})}

\item{package}{Is this a package or a regular project? (Default \code{FALSE})}

\item{rmd}{Should rmarkdown file be used (Default \code{FALSE})}
}
\description{
Add a README.md file to the project directory
}
\seealso{
\code{\link{add_contributing}}, \code{\link{add_license}}, \code{\link{add_license_header}}
}
