% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_readme_rmd}
\alias{use_bcgov_readme_rmd}
\title{Add a README.md file to the project directory}
\usage{
use_bcgov_readme_rmd(project = NULL, package = FALSE)
}
\arguments{
\item{project}{Name of the project. Defaults to the name of the Rstudio project/working directory}

\item{package}{Is this a package or a regular project? (Default  \code{FALSE})}
}
\description{
Add a README.md file to the project directory
}
\seealso{
\code{\link[=use_bcgov_contributing]{use_bcgov_contributing()}}, \code{\link[=use_bcgov_license]{use_bcgov_license()}}, \code{\link[=add_license_header]{add_license_header()}}
}
