% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_contributing}
\alias{add_contributing}
\title{Add a CONTRIBUTING.md file to the project directory}
\usage{
add_contributing(path = ".", package = FALSE)
}
\arguments{
\item{path}{Directory path (default \code{"."})}

\item{package}{Is this a package or a regular project? (Default \code{FALSE}).
If \code{TRUE}, "CONTRIBUTING.md" will be added to .Rbuildignore}
}
\description{
Add a CONTRIBUTING.md file to the project directory
}
\seealso{
\code{\link{add_readme}}, \code{\link{add_license}}, \code{\link{add_license_header}}
}

