% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_rproj}
\alias{add_rproj}
\title{Add an RProject file to a directory}
\usage{
add_rproj(path = ".", outfile)
}
\arguments{
\item{path}{folder path of the project. Default \code{"."}}

\item{outfile}{the name of the RProj file}
}
\description{
Add an RProject file to a directory
}
\seealso{
\code{\link{add_readme}}, \code{\link{add_license}}, \code{\link{add_license_header}}
}
