% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_skeleton.R
\name{analysis_skeleton}
\alias{analysis_skeleton}
\title{Creates the framework of a new analysis development folder}
\usage{
analysis_skeleton(path = ".", git_init = TRUE, git_clone = NULL,
  rstudio = TRUE, apache = TRUE, CoC = TRUE,
  copyright_holder = "Province of British Columbia")
}
\arguments{
\item{path}{location to create new analysis. If \code{"."} (the default), 
the name of the working directory will be taken as the analysis name. If 
not \code{"."}, the last component of the given path will be used as the 
analysis name.}

\item{git_init}{Create a new git repository? Logical, default \code{TRUE}.}

\item{git_clone}{the url of a git repo to clone.}

\item{rstudio}{Create an Rstudio project file?}

\item{apache}{Add licensing info for release under Apache 2.0? Default \code{TRUE}.}

\item{CoC}{Should a Code of Conduct be added to the repository? Default \code{TRUE}.}

\item{copyright_holder}{the name of the copyright holder (default 
"Province of British Columbia). Only necessary if adding a license}
}
\description{
Creates the folder structure for a new analysis.
}
\details{
If you are cloning a repository (\code{git_clone = "path_to_repo"}),
  you should run this function from the root of your dev folder and leave 
  \code{path = "."}, as the repository will be cloned into a new folder. If 
  you are setting up a new project (with or without git), \code{path} should 
  be \code{"."} if you are within an already created project directory, or 
  the name of the folder you want to create.
}
\examples{
\donttest{
 analysis_skeleton(path = "c:/_dev/tarballs", rstudio = TRUE)
}
}
