% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_bcgov_package}
\alias{create_bcgov_package}
\title{Create a bcgov R package directory structure}
\usage{
create_bcgov_package(path = ".", rmarkdown = TRUE,
  coc_email = getOption("bcgovr.coc.email", default = NULL), open = TRUE)
}
\arguments{
\item{path}{Path to the directory in which to initialize the project.
Default \code{"."} - your current working directory.}

\item{rmarkdown}{Should an Rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{TRUE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}
}
\description{
Create a package directory structure for a new bcgov R package in your current working directory.
}
\examples{
\donttest{
 bcgovr::create_bcgov_package()
}
}
