% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_bcgov_package}
\alias{create_bcgov_package}
\title{Create a bcgov R package directory structure}
\usage{
create_bcgov_package(path = ".", rmarkdown = TRUE,
  coc_email = get_coc_email(), rstudio = rstudioapi::isAvailable(),
  open = TRUE)
}
\arguments{
\item{path}{Path to the directory in which to initialize the project.
Default \code{"."} - your current working directory.}

\item{rmarkdown}{Should an Rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{TRUE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct. It is
recommended that you save this setting by adding a line like:
\code{options("bcgovr.coc.email" = "your.email@gov.bc.ca")} to your \code{.Rprofile}.
Use \code{\link[usethis:edit]{usethis::edit_r_profile()}} for an easy way to find and edit this file.}

\item{rstudio}{If \code{TRUE}, calls \code{\link[usethis:use_rstudio]{usethis::use_rstudio()}} to make the new
project into an RStudio Project. If \code{FALSE}, a \code{.here} file is created so
that the directory can be recognized as a project by the \code{here} or \code{rprojroot} packages.}

\item{open}{If TRUE and in RStudio, the new project is opened in a new instance,
if possible, or is switched to, otherwise.}
}
\description{
Create a package directory structure for a new bcgov R package in your current working directory.
}
\examples{
\dontrun{
 bcgovr::create_bcgov_package()
}
}
