% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_from_bcgov_github}
\alias{create_from_bcgov_github}
\title{Create a local repository from a bcgov GitHub repository}
\usage{
create_from_bcgov_github(repo, destdir = ".", protocol = "https", ...)
}
\arguments{
\item{repo}{bcgov GitHub repo name specified like this: \code{bcgov/reponame}}

\item{destdir}{The destination directory where the cloned project will be stored locally}

\item{protocol}{Transfer protocol. One of \code{'https'} (default) or \code{'ssh'}.}

\item{...}{Other arguments passed on to \code{\link[usethis:create_from_github]{usethis::create_from_github()}}}
}
\description{
Creates a new local git repository cloned from a bcgov GitHub repository
}
\examples{
\donttest{
create_from_bcgov_github("bcgov/bcgovr")
}

}
