% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devex_badge.R
\name{devex_badge}
\alias{devex_badge}
\title{Add html for BC DevExchange project state badge}
\usage{
devex_badge(project_state, cat = TRUE)
}
\arguments{
\item{project_state}{one of:'inspiration', 'exploration', or 'delivery'}

\item{cat}{use cat to print the result (\code{TRUE}; default) or return a character vector (\code{FALSE})?}
}
\value{
html
}
\description{
Add html for BC DevExchange project state badge
}
