% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devex_badge.R
\name{insert_bcgov_devex_badge}
\alias{insert_bcgov_devex_badge}
\title{Add html for inserting a BC DevExchange project state badge}
\usage{
insert_bcgov_devex_badge(project_state, cat = TRUE)
}
\arguments{
\item{project_state}{one of: 'inspiration', 'exploration', 'dormant', 'delivery' or 'retired'}

\item{cat}{use cat to print the result (\code{TRUE}; default) or return a character vector (\code{FALSE})?}
}
\value{
html
}
\description{
Add html for inserting a BC DevExchange project state badge
}
