% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devex_badge.R
\name{insert_bcgov_lifecycle_badge}
\alias{insert_bcgov_lifecycle_badge}
\title{Add html for inserting a bcgov lifecycle project state badge}
\usage{
insert_bcgov_lifecycle_badge(project_state, cat = TRUE)
}
\arguments{
\item{project_state}{One of: 'experimental', 'maturing', 'dormant', 'stable' or 'retired'. A project state
badge should be included in a readme to indicate the project lifecycle stage. Recognized project states are
described at \url{https://github.com/bcgov/repomountie/blob/8b2ebdc9756819625a56f7a426c29f99b777ab1d/doc/state-badges.md}}

\item{cat}{Use cat to print the result (\code{TRUE}; default) or return a character vector (\code{FALSE})?}
}
\value{
html
}
\description{
Add html for inserting a bcgov lifecycle project state badge
}
