% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_skeleton.R
\name{package_skeleton}
\alias{package_skeleton}
\title{Creates the framework of a new package development folder}
\usage{
package_skeleton(path = ".", git_init = TRUE, git_clone = NULL,
  apache = TRUE, rstudio = TRUE, CoC = TRUE, rmarkdown = TRUE,
  coc_email = getOption("bcgovr.coc.email"),
  copyright_holder = "Province of British Columbia")
}
\arguments{
\item{path}{location to create new analysis. If \code{"."} (the default), 
the name of the working directory will be taken as the analysis name. If 
not \code{"."}, the last component of the given path will be used as the 
analysis name.}

\item{git_init}{Create a new git repository? Logical, default \code{TRUE}.}

\item{git_clone}{the url of a git repo to clone.}

\item{apache}{Add licensing info for release under Apache 2.0? Default \code{TRUE}.}

\item{rstudio}{Create an Rstudio project file? If true, a new RStudio session will open.}

\item{CoC}{Should a Code of Conduct be added to the repository? Default \code{TRUE}.}

\item{rmarkdown}{Should an rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{TRUE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct. 

You may want to set this option (\code{options("bcgovr.coc.email" = "my.email@gov.bc.ca")}) in your 
.Rprofile file so that every time you start a new analysis, it will be automatically populated.}

\item{copyright_holder}{the name of the copyright holder (default 
"Province of British Columbia). Only necessary if adding a license}
}
\description{
Creates the framework of a new package development folder
}
\details{
If you are cloning a repository (\code{git_clone = "path_to_repo"}),
  you should run this function from the root of your dev folder and leave 
  \code{path = "."}, as the repository will be cloned into a new folder. If 
  you are setting up a new project (with or without git), \code{path} should 
  be \code{"."} if you are within an already created project directory, or 
  the name of the folder you want to create.
}
\examples{
\donttest{
 bcgovr::package_skeleton(path = "c:/_dev/tarballs")
}
}
