% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roxygen_template.R
\name{roxygen_template}
\alias{roxygen_template}
\title{Populate the boilerplate roxygen template at the top of the function.}
\usage{
roxygen_template(funfile, func, export = TRUE)
}
\arguments{
\item{funfile}{path to the .R file containing the function}

\item{func}{the name of the function you want to document}

\item{export}{Is the function exported (default \code{TRUE})? If \code{FALSE}
keyword 'internal' is added}
}
\value{
nothing, but adds the roxygen template to the top of the file
}
\description{
Inspired by Karthik Ram's RTools Sublime Text 2 plugin:
https://github.com/karthik/Rtools
}

