% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_setup.R
\name{set_cran_repo}
\alias{set_cran_repo}
\title{Set the default CRAN repository from which to download packages}
\usage{
set_cran_repo()
}
\value{
TRUE (invisibly)
}
\description{
This is achieved by setting the `repos["CRAN"]` option to
https://cran.rstudio.com in your `.Rprofile` file
}
\seealso{
set_home
}
