% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_setup.R
\name{set_home}
\alias{set_home}
\title{Set HOME environment variable on Windows}
\usage{
set_home()
}
\value{
the path to `HOME`, invisibly
}
\description{
On many Windows computers in an enterprise environment, the
`HOME` environment variable is either unset, or set to a location
on a network drive. R doesn't work well with this setup, so this
function checks what it is set to, and if it's not on the
`C:/` drive, it sets it to the value of `%USERPROFILE%`, usually
`C:/Users/[username]`.
}
\details{
If you have a `.Rprofile` or `.Renviron` file in your existing `HOME`
directory these will be copied to the new location of `HOME`.

This also ensures that the user's personal package library is at
`C:/Users/[your_username]/R/win-library/[R-version]/` if you
haven't already set it via the `R_LIBS` environment variable. It also moves
all of your packages from your old library location to the new one.
}
\seealso{
set_cran_repo
}
