% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_soe.r
\name{theme_soe}
\alias{theme_soe}
\title{Default theme for EnvReportBC graphs and plots}
\usage{
theme_soe(base_size = 12, base_family = "Verdana")
}
\arguments{
\item{base_size}{base font size (default = 12)}

\item{base_family}{base font family (default = Verdana)}
}
\value{
returns a plot theme
}
\description{
Default theme for EnvReportBC graphs and plots
}
\keyword{plotting}
\keyword{theme}
