% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_soe.r
\name{theme_soe}
\alias{theme_soe}
\title{Default theme for EnvReportBC graphs and plots}
\usage{
theme_soe(base_size = 12, base_family = "Verdana", use_sizes = TRUE)
}
\arguments{
\item{base_size}{base font size (default = 12)}

\item{base_family}{base font family (default = Verdana)}

\item{use_sizes}{use relative font sizes (?)}
}
\value{
returns a plot theme
}
\description{
Default theme for EnvReportBC graphs and plots
}
\keyword{plotting}
\keyword{theme}

