% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_soe_facet}
\alias{theme_soe_facet}
\title{Soe plot theme for facetted graphs}
\usage{
theme_soe_facet(base_size = 12, base_family = "Verdana", use_sizes = TRUE)
}
\arguments{
\item{base_size}{base font size (default = 12)}

\item{base_family}{base font family (default = Verdana)}

\item{use_sizes}{use relative font sizes (?)}
}
\value{
a ggplot2 theme
}
\description{
Soe plot theme for facetted graphs
}
\examples{
\dontrun{

}
}
\keyword{plotting}
\keyword{theme}

