% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_code_of_conduct}
\alias{use_bcgov_code_of_conduct}
\title{Add a CODE_OF_CONDUCT.md file to the project directory}
\usage{
use_bcgov_code_of_conduct(coc_email = getOption("bcgovr.coc.email", default =
  coc_stop()))
}
\arguments{
\item{coc_email}{Contact email address(es) for the Code of Conduct. It is
recommended that you save this setting by adding a line like:
\code{options("bcgovr.coc.email" = "your.email@gov.bc.ca")} to yor \code{.Rprofile}.
Use \code{\link[usethis:edit_r_profile]{usethis::edit_r_profile()}} for an easy way to find and edit this file.}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Add a CODE_OF_CONDUCT.md file to the project directory
}
\seealso{
\code{\link[=use_bcgov_readme]{use_bcgov_readme()}}, \code{\link[=use_bcgov_licence]{use_bcgov_licence()}}, \code{\link[=use_bcgov_contributing]{use_bcgov_contributing()}}
}
