% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_git}
\alias{use_bcgov_git}
\title{Initialize a git repository, with necessary bcgov infrastructure}
\usage{
use_bcgov_git(rmarkdown = TRUE, licence = "apache2",
  coc_email = getOption("bcgovr.coc.email", default = NULL),
  message = "Initial commit")
}
\arguments{
\item{rmarkdown}{Should an rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{FALSE}.}

\item{licence}{Which license to apply? Default is Apache 2.0 (\code{"apache2"}).
Use \code{"cc-by"} for \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0}}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}

\item{message}{git commit message}
}
\description{
Initialize a git repository, with necessary bcgov infrastructure
}
