% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_github}
\alias{use_bcgov_github}
\title{Add your project to bcgov GitHub}
\usage{
use_bcgov_github(organisation = "bcgov", rmarkdown = TRUE,
  licence = "apache2", coc_email = get_coc_email(),
  protocol = "https", ...)
}
\arguments{
\item{organisation}{GitHub organisation where the repo will be hosted.
One of \code{'bcgov'} (default), \code{'bcgov-c'}, or \code{NULL} to set up in your
personal GitHub account}

\item{rmarkdown}{Should an Rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{FALSE}.}

\item{licence}{Which licence to apply? Default is Apache 2.0 (\code{"apache2"}).
Use \code{"cc-by"} for \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0}}

\item{coc_email}{Contact email address(es) for the Code of Conduct. It is
recommended that you save this setting by adding a line like:
\code{options("bcgovr.coc.email" = "your.email@gov.bc.ca")} to your \code{.Rprofile}.
Use \code{\link[usethis:edit]{usethis::edit_r_profile()}} for an easy way to find and edit this file.}

\item{protocol}{Transfer protocol. One of \code{'https'} (default) or \code{'ssh'}.}

\item{...}{Other arguments passed on to \code{\link[usethis:use_github]{usethis::use_github()}}}
}
\description{
This function requires that your project already be a Git repository. Use
\code{\link[=use_bcgov_git]{use_bcgov_git()}} to initialise a repository if necessary.
}
