% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_pkgdown}
\alias{use_bcgov_pkgdown}
\title{Create a pkgdown site for your bcgov package, with bcgov styling}
\usage{
use_bcgov_pkgdown()
}
\description{
Create a pkgdown site for your bcgov package, with bcgov styling
}
