% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_pkgdown}
\alias{use_bcgov_pkgdown}
\title{Create a pkgdown site for your bcgov package, with bcgov styling}
\usage{
use_bcgov_pkgdown()
}
\description{
\href{https://usethis.r-lib.org}{pkgdown} makes it easy to turn your package into
a beautiful website. For an example, see https://bcgov.github.io/bcmaps.
}
\details{
This helper creates a \code{pkgdown/} directory with a \code{_pkgdown.yml} file for you
to customize your site, and an \code{extra.css} file which styles your site with
the bcgov colour scheme.

It also adds a \code{docs/} folder for you, and adds all these files to
\code{.Rbuildignore}.
}
