% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_readme}
\alias{use_bcgov_readme}
\title{Add a README.md file to the project directory}
\usage{
use_bcgov_readme(project = NULL, licence = c("apache2", "cc-by"))
}
\arguments{
\item{project}{Name of the project. Defaults to the name of the Rstudio project/working directory}

\item{path}{Directory path (default \code{"."})}
}
\description{
Add a README.md file to the project directory
}
\seealso{
\code{\link[=use_bcgov_contributing]{use_bcgov_contributing()}}, \code{\link[=use_bcgov_licence]{use_bcgov_licence()}}, \code{\link[=use_bcgov_code_of_conduct]{use_bcgov_code_of_conduct()}}
}
