% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_req}
\alias{use_bcgov_req}
\title{Add bcgov GitHub requirements to your project directory}
\usage{
use_bcgov_req(
  rmarkdown = TRUE,
  coc_email = get_coc_email(),
  licence = "apache2"
)
}
\arguments{
\item{rmarkdown}{Should an Rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{FALSE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct. It is
recommended that you save this setting by adding a line like:
\code{options("bcgovr.coc.email" = "your.email@gov.bc.ca")} to your \code{.Rprofile}.
Use \code{\link[usethis:edit]{usethis::edit_r_profile()}} for an easy way to find and edit this file.}

\item{licence}{Which licence to apply? Default is Apache 2.0 (\code{"apache2"}).
Use \code{"cc-by"} for \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0}}
}
\description{
Add a LICENCE file, a README, a CODE OF CONDUCT and a CONTRIBUTING file
}
