% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_req}
\alias{use_bcgov_req}
\title{Add BC Government requirements to your project directory}
\usage{
use_bcgov_req(rmarkdown = TRUE)
}
\arguments{
\item{rmarkdown}{Should an rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{FALSE}.}
}
\description{
Adds a LICENSE file, a README, a CODE OF CONDUCT and a CONTRIBUTING file
}
