% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_bcgov_req}
\alias{use_bcgov_req}
\title{Add bcgov GitHub requirements to your project directory}
\usage{
use_bcgov_req(rmarkdown = TRUE, coc_email = getOption("bcgovr.coc.email",
  default = NULL), licence = "apache2")
}
\arguments{
\item{rmarkdown}{Should an rmarkdown file be added to the repository
with its corresponding markdown file? Default \code{FALSE}.}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}

\item{licence}{Which licence to apply? Default is Apache 2.0 (\code{"apache2"}).
Use \code{"cc-by"} for \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0}}
}
\description{
Add a LICENSE file, a README, a CODE OF CONDUCT and a CONTRIBUTING file
}
